Fixed a problem with long Accept requests
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@718 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/http.cpp
@@ -1363,7 +1363,7 @@ int http::controlConnection(LPCONNECTION a, char* /*b1*/, char* /*b2*/, int bs1,
 		}
 	}
 	
-	u_long validRequest=http_headers::buildHTTPRequestHeaderStruct(&td.request, &td);
+	int validRequest=http_headers::buildHTTPRequestHeaderStruct(&td.request, &td);
 	if(validRequest==(u_long)-1)/*!If the header is incomplete returns 2*/
 	{
 		if(!strcmp(td.request.VER, "HTTP/1.1"))/*Be sure that the client can handle the 100 status code*/
